refactor: move https/travis helper functions to testlab
move https/travis helper functions to testlab
